requirement.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.2
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-export==1.11.0
django-formtools==2.2
django-import-export==2.4.0
django-object-tools==2.0.0
django-reversion==3.0.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
install==1.3.4
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.0
Pygments==2.2.0
PyJWT==1.7.1
pytz==2020.5
PyYAML==5.3.1
qcloudsms-py==0.1.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.2
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==2.0.1
xlwt==1.3.0
